refactor(plan_apply): extract per-terminal handlers to terminal.rs (a)
Move the 390-line match block from spawn_executor_run into three
dedicated handlers (handle_completed, handle_cancelled, handle_paused)
plus a cancel_pending_items helper with retry-once resilience.

spawn_executor_run body reduced to ~50 lines delegating to
super::terminal::handle_* calls. No behavioral change; all 49
plan_apply tests pass.
